requirements.txt
# Krishi-Setu Pipeline Dependencies

requests==2.31.0
sentence-transformers==2.7.0
faiss-cpu==1.7.4
numpy==1.26.4
schedule==1.2.1
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1